Clean up API and implementation.

+++ runtime/src/runtime.c
@@ -120,9 +120,9 @@ fmr_return lf_invoke(struct _lf_module *module, fmr_function function, struct _l
 /* Hacky way to compute the appropriate pointer argument for a device. */
 fmr_va fmr_ptr(struct _lf_device *device, void *ptr) {
 	if (device -> configuration.attributes & lf_device_32bit) {
-		return fmr_int32(ptr);
+		return fmr_int32((uintptr_t)ptr);
 	} else if (device -> configuration.attributes & lf_device_16bit) {
-		return fmr_int16(ptr);
+		return fmr_int16((uintptr_t)ptr);
 	} else {
 		lf_error_raise(E_FMR, error_message("No pointer size specified for the target architecture."));
 	}
